Implemented undoing (i.e. deleting) action records
Along with the titular change, I've also moved event-related stuff into
the event views folder, moved some functionality into a new eventutils
file, and changed settings such that developers in debug mode are not
required to constantly run collectstatic (ideally, production should
still be unchanged).
