+++ frontend/src/Components/Events/eventutils.js
@@ -1,0 +1,43 @@
+async function reactToResponse(response) {
+    let validResponseStatuses = [200, 201, 204];
+    if (validResponseStatuses.some((stat) => response.status == stat)) {
+        // const result = await response.json();
+        window.location.reload();
+    } else {
+        const message = await response.json();
+        alert(`${response.statusText}: ${message['detail']}`);
+    }
+}
+export async function requestAction(event, action, userActedOn) {
+    const response = await fetch(`/api/eventactionrecords/`, {
+        method: "POST",
+        headers: {
+            "Content-Type": "application/json",
+            "X-CSRFToken": document.cookie
+                .split("; ")
+                .find((element) => element.startsWith("csrftoken="))
+                .split("=")[1],
+        },
+        body: JSON.stringify({
+            event: event.pk,
+            acted_on: userActedOn.user_id,
+            action: action,
+            extra_data: "{}",
+            points: action === "Check Off" ? event.points: 0,
+        }),
+    });
+    await reactToResponse(response);
+}
+
+export async function deleteAction(actionId) {
+    const response = await fetch(`/api/eventactionrecords/${actionId}/`, {
+        method: "DELETE",
+        headers: {
+            "X-CSRFToken": document.cookie
+                .split("; ")
+                .find((element) => element.startsWith("csrftoken="))
+                .split("=")[1],
+        }
+    })
+    await reactToResponse(response);
+}
